Move client_session_context to client/session.py
Importing it from client/main.py pulled in the eager OpenAI() client
construction, which broke unit-test collection in CI where
OPENAI_API_KEY isn't set.
