+++ .changes/unreleased/Bug Fix-20260506-090056.yaml
@@ -1,0 +1,3 @@
+kind: Bug Fix
+body: Move client_session_context out of client/main.py so importing it in tests no longer requires OPENAI_API_KEY.
+time: 2026-05-06T09:00:56.84291-05:00
