+++ src/client/main.py
@@ -1,20 +1,13 @@
 import asyncio
 import json
-from collections.abc import AsyncGenerator
-from contextlib import asynccontextmanager
 from time import time
 
-from mcp.client.session import ClientSession
-from mcp.shared.memory import (
-    create_connected_server_and_client_session as client_session,
-)
 from openai import OpenAI
 from openai.types.responses.response_input_param import FunctionCallOutput
 from openai.types.responses.response_output_message import ResponseOutputMessage
 
+from client.session import client_session_context
 from client.tools import map_tools
-from dbt_mcp.config.config import load_config
-from dbt_mcp.mcp.server import create_dbt_mcp
 
 LLM_MODEL = "gpt-4o-mini"
 TOOL_RESPONSE_TRUNCATION = 100  # set to None for no truncation
@@ -22,14 +15,6 @@ TOOL_RESPONSE_TRUNCATION = 100  # set to None for no truncation
 
 llm_client = OpenAI()
 messages = []
-
-
-@asynccontextmanager
-async def client_session_context() -> AsyncGenerator[ClientSession, None]:
-    config = load_config()
-    server = await create_dbt_mcp(config)
-    async with client_session(server) as client:
-        yield client
 
 
 async def main():
